+++ Gp.Domain/Models/DespesaLancamento.cs
@@ -5,7 +5,7 @@ namespace Gp.Domain.Models
     {
         public string Descricao { get; set; }
         public decimal ValorTotal { get; set; }
-        public int DespesaId { get; set; }
+        public long DespesaId { get; set; }
         public virtual Despesa Despesa { get; set; }
     }
 }
