+++ Gp.Domain/Models/ReceitaLancamento.cs
@@ -4,7 +4,7 @@
     {
         public string Descricao { get; set; }
         public decimal ValorTotal { get; set; }
-        public int ReceitaId { get; set; }
+        public long ReceitaId { get; set; }
         public virtual Receita Receita { get; set; }
     }
 }
